do not allow to subscribe to closed channels
Show a message that the user can't subscribe to the channel if `channel.subscribeable==False`.
Do not show the subscription form.

Currently channel.subscribeable only affects if the channel is listed in subscription forms, but the naming suggest, that setting this to False would not allow users to subscribe to this channel at all.